move global cors config
